Store user in a ref and use this to check login and remove the explicit timeout for the data
If I understand correctly, it is enough to check that the reference contains
a user to validate if the user is connected or not. Therefore there is no need
to check the password in the database each time. This reduces the number of
call to the database. To be sure that the user is correct, we use a group
session, so this reference is share for all session of a same user.

Use the default value for timeout for the reference from eliom.